World Settings panel + UI reorg; grid; code-review fixes L2/L4-L7/L12/L14/L15
- World Settings (per-world, saved in map): entry/exit through-traffic
  enable/multiplier/base-rate/rush-hour; MapSerializer v5; unpack sync
- Through-traffic now runs without homes (moved above schedule-mode gate)
- Menu bar: Visibility + World submenus; legend moved into pause menu
- Selection panel hidden when nothing selected; zoom-adaptive higher-contrast grid
- Reseed sim RNG on built-in world load; remove V/[ ]/Sh+Click/Ctrl+O hotkeys
- Code-review: L2 side-array lifecycle, L4 reroute closed-edge, L5 edge-count,
  L6 EdgeStartIdx, L7 flags ushort, L12 diag gates, L14 ground-art cull, L15 curve AABB

+++ Roads.App/World/RoadNode.cs
@@ -61,8 +61,11 @@ public struct RoadNode
 {
     /// <summary>World-space position in meters.</summary>
     public Vector2 Position;
-    /// <summary>Start index into RoadGraph's adjacency array for this node's outgoing edges.</summary>
-    public ushort EdgeStartIdx;
+    /// <summary>Start index into RoadGraph's adjacency array for this node's outgoing edges.
+    /// An int: the index counts TOTAL active edges graph-wide, and a ushort silently wrapped
+    /// past 65,535 (≈33K two-way roads — within city-scale ambitions), making nodes read
+    /// other nodes' outgoing lists. Not serialized — rebuilt by RebuildAdjacency.</summary>
+    public int EdgeStartIdx;
     /// <summary>Number of outgoing edges from this node.</summary>
     public byte EdgeCount;
     /// <summary>Traffic control flags for this intersection.</summary>
